requirements.txt
opencv-python == 4.8.1
Pillow == 9.0.1
scipy == 1.11.3